Nail-down GCE spin-up logic
Utilize a gcloud sdk container to run all operations. This is really not
necessary for CI BUT in order to provide a uniform testing environment
against local development and CI this was necessary. Otherwise if a
local developer used their machine loaded gcloud sdk, there is a
possiblility of collusion of permissions and things launching with
elevated permissions. This is not necessarilly a dangerous situation
just one that doesnt adequately mimic CI enough...

So tl;dr Lets make CI and local dev CI testing as close as possible